refactor `__init__.py`
